createJobStatus moved, some namings changed, comments added
