+++ app/config/config.yml
@@ -174,13 +174,13 @@ old_sound_rabbit_mq:
             heartbeat: 0
     producers:
         job:
-            class: Graviton\MessageBundle\Service\JobProducer
+            class: Graviton\MessageBundle\Service\MessageProducer
             connection: default
             exchange_options:
                 name: test-topic
                 type: topic
     consumers:
-        test:
+        dump:
             connection: default
             exchange_options:
                 name: test-topic
@@ -188,4 +188,4 @@ old_sound_rabbit_mq:
             queue_options:
                 routing_keys:
                     - core.app.*
-            callback:         test_consumer
+            callback: graviton.message.consumer.dump
