--- src/Graviton/MessageBundle/Repository/JobStatusRepository.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Graviton\MessageBundle\Repository;
-
-use Doctrine\ODM\MongoDB\DocumentRepository;
-
-class JobStatusRepository extends DocumentRepository
-{
-
-}
