+++ src/Graviton/MessageBundle/Resources/config/services.xml
@@ -3,16 +3,13 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
-        <parameter key="graviton.message.doctrineeventpublisher.class">Graviton\MessageBundle\Service\DoctrineEventPublisher</parameter>
-        <parameter key="graviton.message.repository.job.class">Graviton\MessageBundle\Repository\JobRepository</parameter>
+        <parameter key="graviton.message.doctrineeventpublisher.class">Graviton\MessageBundle\Service\DocumentEventPublisher</parameter>
         <parameter key="graviton.rest.routing.collection.class">Symfony\Component\Routing\RouteCollection</parameter>
     </parameters>
     <services>
+        <service id="graviton.message.consumer.dump" class="Graviton\MessageBundle\Consumer\DumpConsumer" />
         <service id="graviton.message.jobproducer" parent="old_sound_rabbit_mq.job_producer">
             <property name="registeredRoutingKeys[]">core.app.create</property>
-            <call method="setJobRepository">
-                <argument type="service" id="service_container" />
-            </call>
         </service>
         <service id="graviton.message.doctrineeventpublisher" class="%graviton.message.doctrineeventpublisher.class%">
             <argument type="service" id="graviton.message.jobproducer" />
@@ -21,16 +18,8 @@
             <tag name="doctrine_mongodb.odm.event_subscriber" />
             <property name="documents['Graviton\CoreBundle\Document\App']">graviton.core.rest.app.get</property>
         </service>
-        <service id="graviton.message.repository.job"
-             class="%graviton.message.repository.job.class%"
-             factory-service="doctrine_mongodb.odm.default_document_manager"
-             factory-method="getRepository">
-            <argument type="string">GravitonMessageBundle:JobStatus</argument>
-        </service>
-
 
         <service id="graviton.rest.routing.collection" class="%graviton.rest.routing.collection.class%"/>
-        <!-- Routing loader -->
         <service id="graviton.rest.routing.loader" class="%graviton.rest.routing.loader.class%">
             <argument type="service" id="graviton.rest.routing.collection"/>
             <argument>%graviton.rest.services%</argument>
